Use '-Dorg.aspectj.weaver.openarchives=20' to further #125
This commit is not meant to be merged into the main branch, but used to
trigger the complete test suite with the lowest possible limit for the
open JAR file AspectJ introduced in JDT Core. As some changes in
eclipse/aspectj.eclipse.jdt.core#19 change behaviour (throwing
exceptions where before they were just logged and dummy results
returned), we want to find out if this breaks any existing tests when
provoking cache exhaustion, i.e. closing and re-opening of JARs.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>