Obsolete workaround for MCOMPILER-485
This commit was meant to work around a supposed problem with Javac,
which actually was caused by a bogus piece Maven Compiler code using ASM
in order to create class files instead of delegating to Javac or ECJ.

Consequently, 'package-info' files without runtime-scoped annotations
would then contain package names with backslash characters when
compiled by Maven Compiler Plugin 3.10.0 on Windows. This bug was fixed
in 3.10.1. Therefore, the workaround in this commit was unnecessary to
begin with.

Relates to #176.

See https://issues.apache.org/jira/browse/MCOMPILER-485.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>